chore: move attr-to-prop helpers to @lwc/shared
