+++ packages/@lwc/template-compiler/src/parser/constants.ts
@@ -4,11 +4,14 @@
  * SPDX-License-Identifier: MIT
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
-import { AriaAttrNameToPropNameMap } from '@lwc/shared';
 import { ElementDirectiveName } from '../shared/types';
 import { HTML_ATTRIBUTE_ELEMENT_MAP } from './utils/html-element-attributes';
 import { HTML_ELEMENTS } from './utils/html-elements';
 import { SVG_ELEMENTS } from './utils/svg-elements';
+
+// Originally defined here, so this re-export is just for backwards compatibility
+// Also, it had a typo!
+export { ATTRS_PROPS_TRANSFORMS as ATTRS_PROPS_TRANFORMS } from '@lwc/shared';
 
 export const EXPRESSION_RE = /(\{(?:.)+?\})/g;
 
@@ -105,26 +108,6 @@ export const SUPPORTED_SVG_TAGS = new Set([
 ]);
 
 export const DISALLOWED_MATHML_TAGS = new Set(['script', 'link', 'base', 'object']);
-
-export const ATTRS_PROPS_TRANFORMS: { [name: string]: string } = {
-    accesskey: 'accessKey',
-    readonly: 'readOnly',
-    tabindex: 'tabIndex',
-    bgcolor: 'bgColor',
-    colspan: 'colSpan',
-    rowspan: 'rowSpan',
-    contenteditable: 'contentEditable',
-    crossorigin: 'crossOrigin',
-    datetime: 'dateTime',
-    formaction: 'formAction',
-    ismap: 'isMap',
-    maxlength: 'maxLength',
-    minlength: 'minLength',
-    novalidate: 'noValidate',
-    usemap: 'useMap',
-    for: 'htmlFor',
-    ...AriaAttrNameToPropNameMap,
-};
 
 export const DISALLOWED_HTML_TAGS = new Set(['base', 'link', 'meta', 'script', 'title']);
