+++ packages/@lwc/template-compiler/src/parser/index.ts
@@ -8,6 +8,7 @@ import {
     HTML_NAMESPACE,
     SVG_NAMESPACE,
     MATHML_NAMESPACE,
+    attributeToPropertyName,
     isVoidElement,
     isUndefined,
     isNull,
@@ -61,7 +62,6 @@ import {
 } from './expression';
 import {
     attributeName,
-    attributeToPropertyName,
     isAttribute,
     isProhibitedIsAttribute,
     isTabIndexAttribute,
